+++ src/chains/Sonic/Common.sol
@@ -16,7 +16,14 @@ import {
     uniswapV3ForkId,
     IUniswapV3Callback
 } from "../../core/univ3forks/UniswapV3.sol";
+import {
+    solidlyV3SonicFactory,
+    solidlyV3InitHash,
+    solidlyV3ForkId,
+    ISolidlyV3Callback
+} from "../../core/univ3forks/SolidlyV3.sol";
 import {spookySwapFactory, spookySwapForkId} from "../../core/univ3forks/SpookySwap.sol";
+import {wagmiFactory, wagmiInitHash, wagmiForkId} from "../../core/univ3forks/Wagmi.sol";
 
 // Solidity inheritance is stupid
 import {SettlerAbstract} from "../../SettlerAbstract.sol";
@@ -51,9 +58,17 @@ abstract contract SonicMixin is FreeMemory, SettlerBase {
             factory = uniswapV3SonicFactory;
             initHash = uniswapV3InitHash;
             callbackSelector = uint32(IUniswapV3Callback.uniswapV3SwapCallback.selector);
+        } else if (forkId == solidlyV3ForkId) {
+            factory = solidlyV3SonicFactory;
+            initHash = solidlyV3InitHash;
+            callbackSelector = uint32(ISolidlyV3Callback.solidlyV3SwapCallback.selector);
         } else if (forkId == spookySwapForkId) {
             factory = spookySwapFactory;
             initHash = uniswapV3InitHash;
+            callbackSelector = uint32(IUniswapV3Callback.uniswapV3SwapCallback.selector);
+        } else if (forkId == wagmiForkId) {
+            factory = wagmiFactory;
+            initHash = wagmiInitHash;
             callbackSelector = uint32(IUniswapV3Callback.uniswapV3SwapCallback.selector);
         } else {
             revert UnknownForkId(forkId);